+++ src/Curry/Base/Position.hs
@@ -67,7 +67,12 @@ instance Read Position where
     [ (Position "" i j noRef, s') | ((i, j), s') <- readsPrec p s ]
 
 instance Show Position where
-  showsPrec _ = showString . show . ppPosition
+  showsPrec _ (Position f l c _) = showString "(Position "
+                                 . shows f . showChar ' '
+                                 . shows l . showChar ' '
+                                 . shows c . showChar ')'
+  showsPrec _ (AST            _) = showString "NoPos"
+  showsPrec _ NoPos              = showString "NoPos"
 
 instance HasPosition Position where
   getPosition = id