+++ src/Curry/Syntax/Pretty.hs
@@ -105,12 +105,16 @@ ppDecl (ExternalDecl   _ fs) = list (map ppIdent fs) <+> text "external"
 ppDecl (PatternDecl _ t rhs) = ppRule (ppPattern 0 t) equals rhs
 ppDecl (FreeDecl       _ vs) = list (map ppIdent vs) <+> text "free"
 
-ppPrec :: Infix -> Integer -> Doc
+ppPrec :: Infix -> Maybe Precedence -> Doc
 ppPrec fix p = ppAssoc fix <+> ppPrio p
-  where ppAssoc InfixL = text "infixl"
-        ppAssoc InfixR = text "infixr"
-        ppAssoc Infix  = text "infix"
-        ppPrio p' = if p' < 0 then empty else integer p'
+  where
+    ppPrio Nothing   = empty
+    ppPrio (Just p') = integer p'
+
+ppAssoc :: Infix -> Doc
+ppAssoc InfixL = text "infixl"
+ppAssoc InfixR = text "infixr"
+ppAssoc Infix  = text "infix"
 
 ppTypeDeclLhs :: String -> Ident -> [Ident] -> Doc
 ppTypeDeclLhs kw tc tvs = text kw <+> ppIdent tc <+> hsep (map ppIdent tvs)